# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRMAI - Full Featured CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List accounts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountList"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # OPPORTUNITIES
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List cases
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: contactId
          in: query
          schema:
            type: integer
        - name: accountId
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseList"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseInput"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseInput"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # USERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
      required:
        - success
        - id

    # USER
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        employees:
          type: integer
          nullable: true
        annualRevenue:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        stage:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        closeDate:
          type: string
          nullable: true
        probability:
          type: integer
          nullable: true
        forecastCategory:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        optyOwner:
          type: string
          nullable: true
        optyTeam:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        createdBy:
          type: integer
          nullable: true
        modifiedBy:
          type: integer
          nullable: true
        clmEnabled:
          type: boolean
        ownerName:
          type: string
          nullable: true
        contactCount:
          type: integer
        dealCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - contactCount
        - dealCount
        - createdAt
        - updatedAt
    AccountList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateAccountInput:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        employees:
          type: integer
          nullable: true
        annualRevenue:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        stage:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        closeDate:
          type: string
          nullable: true
        probability:
          type: integer
          nullable: true
        forecastCategory:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        optyOwner:
          type: string
          nullable: true
        optyTeam:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        clmEnabled:
          type: boolean
      required:
        - name
    UpdateAccountInput:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        employees:
          type: integer
          nullable: true
        annualRevenue:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        stage:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        closeDate:
          type: string
          nullable: true
        probability:
          type: integer
          nullable: true
        forecastCategory:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        optyOwner:
          type: string
          nullable: true
        optyTeam:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        clmEnabled:
          type: boolean

    # CONTACT
...
    Case:
      type: object
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        subject:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, in_progress, pending, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        accountId:
          type: integer
          nullable: true
        accountName:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        resolution:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - caseNumber
        - subject
        - status
        - priority
        - createdAt
        - updatedAt
    CaseList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Case"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateCaseInput:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, in_progress, pending, resolved, closed]
          default: new
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        type:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        contactId:
          type: integer
          nullable: true
        accountId:
          type: integer
          nullable: true
        assignedTo:
          type: integer
          nullable: true
      required:
        - subject
    UpdateCaseInput:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, in_progress, pending, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        contactId:
          type: integer
          nullable: true
        accountId:
          type: integer
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        resolution:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true

    # QUOTE
...